Abandoning the notion of 'observer context'...
As used currently 'observer context' is a call router for 'log' and 'error' calls that may occur outside the normal call scope of app logic (e.g. at load time, or in response to a button click, request completion etc. etc.

The naming 'observer context' is problematic however. So I've renamed the concept 'backchannel' and promoted it to ONMjs-core _even though core doesn't use it at all_... The reason to stuff it in core is because it's tiny, and because I want it available to other library and app code w/only a dependency on ONMjs core.

'backchannel' is suggestive of the purpose, backchannel.log backchannel.error are easy to type and concise.